==> brain_connect_and_si_scripts/8_siq_102020.Rmd <==
---
title: "siq"
author: "Jackie"
date: "10/29/2020"
output: html_notebook
---

Data structure: SI information based on the Suicidal Ideation Questionnaire-JR (Reynolds, 1987) 
- 15 items ranging from 0 - 6 (higher score = higher severity)
- For total, sum scores for 15 items, for total possible score of 90


# Loading libraries
```{r, message=FALSE}
library(tidyverse)
library(haven)
library(modelr)
library(readxl)
library(Rcpp)
library(httpuv)
library(corrr)
library(seriation)
#library(d3heatmap)
library(heatmaply)
library(corrplot)
library(haven)
library(naniar)
library(brms)
library(rstan)
library(bayesplot)
library(rstanarm)
library(knitr)
library(loo)
library(nonnest2)
```

# filepaths
```{r, echo=FALSE, message=FALSE}
siq_t3_fp <- "~/Box/Mooddata_Coordinating/1_Lab_Coordinating/Users/JackieSchwartz/PHIND_RS_and_Suicide/Octo2020Analysis/siq.T3.csv" # pulled my directory 10.29.20
graph_fp <- "~/Box/Mooddata_Coordinating/1_Lab_Coordinating/Users/JackieSchwartz/PHIND_RS_and_Suicide/Octo2020Analysis/graph_all.csv"
```

# reading in files
```{r, echo=FALSE, message=FALSE}
siq_t3 <- read_csv(siq_t3_fp)  %>%
  mutate(ELS_ID = factor(ELS_ID))

graph <- read_csv(graph_fp) %>%
  mutate(ELS_ID = factor(ELS_ID))
```

# score siq data

### filtering out duplicated IDs 
_I'm using the .5/.2 sessions bc they are closest to the scan date_
```{r}
siq_t3_dup <- siq_t3 %>%
  filter(duplicated(ELS_ID)
         ) # duplicated IDs
dupIDs <- siq_t3_dup$ELS_ID

siq_t3_clean <- siq_t3 %>%
  mutate(reassessed =
           ifelse(is.na(Session_Type),
                  "no",
                  "yes"
           )
  )

siq_t3_clean_unique <-
  siq_t3_clean %>%
  mutate(
    dup = ifelse(ELS_ID %in% dupIDs, "dup", "0"),
    drop = ifelse(reassessed == "yes" & dup == "dup", "drop", "0")
  ) %>%
  filter(drop == "0")

```

## recoding 999s and 888s
```{r}
vars <- c("siq_1.T3", "siq_2.T3", "siq_3.T3",  "siq_4.T3", "siq_5.T3", "siq_6.T3", "siq_7.T3", "siq_8.T3", "siq_9.T3", "siq_10.T3", "siq_11.T3", "siq_12.T3", "siq_13.T3", "siq_14.T3", "siq_15.T3")

siq_t3_clean_unique <- 
  siq_t3_clean_unique %>%
  replace_with_na_at(.,
                     .vars = vars,
                     condition = ~.x == 999) %>%
  dplyr::select(-reassessed, -dup, -drop)

```

## scoring siq total
```{r}
siq_t3_clean_total <-
  siq_t3_clean_unique %>%
  mutate(
    siq_total.T3 = 
      rowSums(siq_t3_clean_unique[,(which(names(siq_t3_clean_unique)=='siq_1.T3'):which(names(siq_t3_clean_unique)=='siq_15.T3'))])
  ) %>%
  drop_na(siq_total.T3) %>%
  mutate(ELS_ID = factor(ELS_ID))


# rowSums(df[,which(names(df) == 'siq_1.T3'):which(names(df)=='siq_15.T3')])

```


```{r}
siq_t3_clean_total_fp <- "~/Box/Mooddata_Coordinating/1_Lab_Coordinating/Users/JackieSchwartz/PHIND_RS_and_Suicide/Octo2020Analysis/siq_t3_response.csv"
write_csv(siq_t3_clean_total, siq_t3_clean_total_fp)
```
