==> brain_connect_and_si_scripts/10_glmnet_siq_nest_resamp.Rmd <==
---
title: "10_glmnet_siq_nest_resamp"
author: "Jackie"
date: "4/30/2021; 5/10/21"
output: html_notebook
---


This script performs nested resampling using loocv on inner and outer resamples to obtain relavtively unbiased performance estimation

# loading libraries
```{r, message=FALSE}
library(glmnet)
library(lars)
library(tidyverse)
library(readxl)
library(pscl)
library(MASS)
library(stringr)
library(sjPlot)
library(directlabels)
library(paradox)
library(mlr)

# set seed
set.seed(1234)

# expanding print
options(max.print = 1275)
```

# reading in data
```{r, message=FALSE}
graph_net_fp <- "~/Box/Mooddata_Coordinating/1_Lab_Coordinating/Users/JackieSchwartz/PHIND_RS_and_Suicide/April2021Analysis/fmaps_and_nofmaps/siq_graph.csv"

graph_net <- read_csv(graph_net_fp)
```


# function to z-score
```{r}
options(scipen=999) 
z_score <- function(x) {
diff_mu <- x - mean(x, na.rm = T)
sd <- sd(x, na.rm = T)
diff_mu / sd
}
```

## de-selecting variables
```{r}
graph_net_select_brain <- graph_net %>%
  dplyr::select(
    -c(meds_baseline)
    )
# already filtered out those with history of stbs
```

#### recoding and standardizing
```{r}
graph_net_stand_brain <- graph_net_select_brain %>%
  # converting race/ethinicity to factor and binarizing for analysis
  mutate(
    KSADS_Child_Race_by_P.T1 = factor(KSADS_Child_Race_by_P.T1),
    KSADS_Child_Race_by_P.T1_bin = ifelse(KSADS_Child_Race_by_P.T1 == "1", "1", "0"),
    # converting sex and sleepiness to factor
    sex_child = factor(sex_child),
    sex_child = ifelse(sex_child == "2", "1", "0"),
    sleepy = factor(sleepy),
    fmap_applied = factor(fmap_applied)
    ) %>%
  dplyr::select(-KSADS_Child_Race_by_P.T1, -siq_total.T3) %>%
  # z-scoring all numeric variables
  mutate_at(vars(-ELS_ID, -siq_total_log, -KSADS_Child_Race_by_P.T1_bin, -sex_child, -sleepy, -fmap_applied), z_score) %>%
  # arranging with ELS_ID first, then siq, then everything else
  mutate(siq_total_log_z = z_score(siq_total_log)) %>%
  dplyr::select(ELS_ID, siq_total_log_z, siq_total_log, everything())

# mean(graph_net_stand_brain$fmap_applied == "1")

graph_net_stand_fp <- "~/Box/Mooddata_Coordinating/1_Lab_Coordinating/Users/JackieSchwartz/PHIND_RS_and_Suicide/April2021Analysis/fmaps_and_nofmaps/siq_graph_net_stand_brainmod.csv"

graph_net <- write_csv(graph_net_stand_brain, graph_net_stand_fp)
```

#### setting contrasts
```{r}
graph_net_stand_brain <-
  graph_net_stand_brain %>%
  mutate_at(vars(KSADS_Child_Race_by_P.T1_bin, sex_child, sleepy, fmap_applied), factor)
contrasts(graph_net_stand_brain$KSADS_Child_Race_by_P.T1_bin) = contr.treatment(2)
contrasts(graph_net_stand_brain$sex_child) = contr.treatment(2)
contrasts(graph_net_stand_brain$sleepy) = contr.treatment(2)
contrasts(graph_net_stand_brain$fmap_applied) = contr.treatment(2)
```

#### constructing x matrix and y vector
```{r}
graph_net_stand_siq_brain <- graph_net_stand_brain %>%
  dplyr::select(-ELS_ID, -siq_total_log) %>%
  dplyr::select(siq_total_log_z, everything()) # response (dv) variable in first column

#as.matrix
predictors_scaled_brain <- data.matrix(graph_net_stand_siq_brain[2:1271]) # columns 2 - 1271 are predictors
siq_scaled_brain <- graph_net_stand_siq_brain$siq_total_log_z # my response variable
save(predictors_scaled_brain, file = "predictors_scaled_brain.rda")
save(siq_scaled_brain, file = "siq_scaled_brain.rda")
```

## to get sense of coefficient path
```{r}
lasso_siq_brain <- glmnet(predictors_scaled_brain, siq_scaled_brain, family = "gaussian", alpha = 1, standardize = FALSE)
# 
plot(lasso_siq_brain, xvar = "dev", label = TRUE)
```
# nested cv
```{r}
library(remotes)
n <- 106

## Your procedure

cvfit <- cv.glmnet(predictors_scaled_brain, siq_scaled_brain, standardize = FALSE, alpha = 1, nfolds = n)
betas <- coef(cvfit, s = "lambda.min")

# computing rsquared of unnested cv
# supp_siq_scaled is my y
X <- model.matrix(siq_total_log_z ~ ., data = graph_net_stand_siq_brain)[,-1]
# predict
y_hat <- predict(cvfit,  newx = X,  s = "lambda.min")
# R-Squared
r_sq <- 1 - (sum((siq_scaled_brain - y_hat)^2) / sum((siq_scaled_brain - mean(siq_scaled_brain))^2))
r_sq # 0.6775243 standardized prior to glmnet
cvfit$lambda.min # 0.1293916

## Get measure of performance of procedure
preds <- double(n)
ybars <- double(n)
for(i in 1:n){
    cvfit.inner <- cv.glmnet(predictors_scaled_brain[-i,], siq_scaled_brain[-i], standardize = FALSE, nfolds = n-1, grouped=FALSE, alpha=1)
    preds[i] <- predict(cvfit.inner, newx = predictors_scaled_brain[i,,drop=FALSE], s = "lambda.min")
    ybars[i] <- mean(siq_scaled_brain[-i])
    cat(i)
}

# total SS
mse0 <- mean((siq_scaled_brain-ybars)^2) # 1.009524
# mean square error
mse <- mean((siq_scaled_brain-preds)^2) # 0.4562436

R2 <-1- mse/mse0 # 0.5480619
```

# zero-order correlations

```{r}
graph_net_corr <- graph_net_select_brain %>%
  # converting race/ethinicity to factor and binarizing for analysis
  mutate(
    KSADS_Child_Race_by_P.T1 = factor(KSADS_Child_Race_by_P.T1),
    KSADS_Child_Race_by_P.T1_bin = ifelse(KSADS_Child_Race_by_P.T1 == "1", "1", "0"),
    # converting sex and sleepiness to factor
    sex_child = factor(sex_child),
    sex_child = ifelse(sex_child == "2", "1", "0"),
    sleepy = factor(sleepy),
    fmap_applied = factor(fmap_applied)
    ) %>%
  dplyr::select(ELS_ID, siq_total_log, everything())
graph_net_corr <-
  graph_net_corr %>%
  mutate_at(vars(KSADS_Child_Race_by_P.T1_bin, sex_child, sleepy, fmap_applied), factor)
contrasts(graph_net_corr$KSADS_Child_Race_by_P.T1_bin) = contr.treatment(2)
contrasts(graph_net_corr$sex_child) = contr.treatment(2)
contrasts(graph_net_corr$sleepy) = contr.treatment(2)
contrasts(graph_net_corr$fmap_applied) = contr.treatment(2)
```

```{r}
labeltheme <- theme(axis.title.y = element_blank(),
                    text = element_text(size = 7))
# internalizing at follow-up
cor.test(graph_net_corr$siq_total_log, graph_net_corr$T3_ysr_internalizing_sum)
cor.test(graph_net_corr$siq_total_log, graph_net_corr$T3_ysr_internalizing_sum, method = "spearman")
internalizing_cor <-
  graph_net_corr %>%
  ggplot(
    aes(
      x = T3_ysr_internalizing_sum, 
      y = siq_total_log
      )
    ) +
  geom_point() +
  geom_smooth(method = "lm") +
  theme_classic() +
  labs(
    x = "Concurrent Internalizing Severity"
  ) +
  labeltheme
internalizing_cor

# externalizing at follow-up
cor.test(graph_net_corr$siq_total_log, graph_net_corr$T3_ysr_externalizing_sum)
cor.test(graph_net_corr$siq_total_log, graph_net_corr$T3_ysr_externalizing_sum, method = "spearman")
externalizing_cor <-
  graph_net_corr %>%
  ggplot(
    aes(
      x = T3_ysr_externalizing_sum, 
      y = siq_total_log
      )
    ) +
  geom_point() +
  geom_smooth(method = "lm") +
  theme_classic() +
  labs(
    x = "Concurrent Externalizing Severity"
  ) +
  labeltheme
externalizing_cor

# within-mod degree of the left lateral superior frontal gyrus (lSFG_L_5.wm_z)
cor.test(graph_net_corr$siq_total_log, graph_net_corr$lSFG_L_5.wm_z)
cor.test(graph_net_corr$siq_total_log, graph_net_corr$lSFG_L_5.wm_z, method = "spearman")

lSFG_L_5.wm_z_cor <-
  graph_net_corr %>%
  ggplot(
    aes(
      x = lSFG_L_5.wm_z, 
      y = siq_total_log
      )
    ) +
  geom_point() +
  geom_smooth(method = "lm") +
  theme_classic() +
  labs(
    x = "Lateral SFG (L) - within-module degree"
  ) +
  labeltheme
lSFG_L_5.wm_z_cor

# participation coefficient of the right rostral middle temporal lobe (rostMTG_R_84.pc)
cor.test(graph_net_corr$siq_total_log, graph_net_corr$rostMTG_R_84.pc)
cor.test(graph_net_corr$siq_total_log, graph_net_corr$rostMTG_R_84.pc, method = "spearman")

rostMTG_R_84.pc_cor <-
  graph_net_corr %>%
  ggplot(
    aes(
      x = rostMTG_R_84.pc, 
      y = siq_total_log
      )
    ) +
  geom_point() +
  geom_smooth(method = "lm") +
  theme_classic() +
  labs(
    x = "Rostral MTL (R) - participation coefficient"
  ) +
  labeltheme
rostMTG_R_84.pc_cor

# participation coefficient of the right postcentral ssuperior parietal lobe (postcentSPL_R_132.pc)
cor.test(graph_net_corr$siq_total_log, graph_net_corr$postcentSPL_R_132.pc)
cor.test(graph_net_corr$siq_total_log, graph_net_corr$postcentSPL_R_132.pc, method = "spearman")
postcentSPL_R_132.pc_cor <-
  graph_net_corr %>%
  ggplot(
    aes(
      x = postcentSPL_R_132.pc, 
      y = siq_total_log
      )
    ) +
  geom_point() +
  geom_smooth(method = "lm") +
  theme_classic() +
  labs(
    x = "Postcentral SPL (R) - participation coefficient"
  ) +
  labeltheme
postcentSPL_R_132.pc_cor 

# participation coefficient of the right opercular area of the inferior frontal gyrus (operIFG_R_38.pc)
cor.test(graph_net_corr$siq_total_log, graph_net_corr$operIFG_R_38.pc)
cor.test(graph_net_corr$siq_total_log, graph_net_corr$operIFG_R_38.pc, method = "spearman")

operIFG_R_38.pc_cor <-
  graph_net_corr %>%
  ggplot(
    aes(
      x = operIFG_R_38.pc, 
      y = siq_total_log
      )
    ) +
  geom_point() +
  geom_smooth(method = "lm") +
  theme_classic() +
  labs(
    x = "IFG (R) - participation coefficient"
  ) +
  labeltheme
operIFG_R_38.pc_cor 

# degree of the left caudal middle temproal lobe (caudMTG_L_81.deg)
cor.test(graph_net_corr$siq_total_log, graph_net_corr$caudMTG_L_81.deg)
cor.test(graph_net_corr$siq_total_log, graph_net_corr$caudMTG_L_81.deg, method = "spearman")
caudMTG_L_81.deg_cor <-
  graph_net_corr %>%
  ggplot(
    aes(
      x = caudMTG_L_81.deg, 
      y = siq_total_log
      )
    ) +
  geom_point() +
  geom_smooth(method = "lm") +
  theme_classic() +
  labs(
    x = "Caudal MTL (L) - degree"
  ) +
  labeltheme
caudMTG_L_81.deg_cor 

# participation coeff of the right medial superior temporal lobe (medSTG_R_70.pc)
cor.test(graph_net_corr$siq_total_log, graph_net_corr$medSTG_R_70.pc)
cor.test(graph_net_corr$siq_total_log, graph_net_corr$medSTG_R_70.pc, method = "spearman")

medSTG_R_70.pc_cor <-
  graph_net_corr %>%
  ggplot(
    aes(
      x = medSTG_R_70.pc, 
      y = siq_total_log
      )
    ) +
  geom_point() +
  geom_smooth(method = "lm") +
  theme_classic() +
  labs(
    x = "Medial STG (R) - participation coefficient"
  ) +
  labeltheme
medSTG_R_70.pc_cor 

# within-mod degree of left orbital area 12/47 (orbOrG_L_43.wm_z)
cor.test(graph_net_corr$siq_total_log, graph_net_corr$orbOrG_L_43.wm_z)
cor.test(graph_net_corr$siq_total_log, graph_net_corr$orbOrG_L_43.wm_z, method = "spearman")
orbOrG_L_43.wm_z_cor <-
  graph_net_corr %>%
  ggplot(
    aes(
      x = orbOrG_L_43.wm_z, 
      y = siq_total_log
      )
    ) +
  geom_point() +
  geom_smooth(method = "lm") +
  theme_classic() +
  labs(
    x = "OrG 12/47 (L) - within-module degree"
  ) +
  labeltheme
orbOrG_L_43.wm_z_cor 

# local efficiency of the left Ventral granular insular gyrus (vldINS_L_169.eff_loc )
cor.test(graph_net_corr$siq_total_log, graph_net_corr$vldINS_L_169.eff_loc)
cor.test(graph_net_corr$siq_total_log, graph_net_corr$orbOrG_L_43.wm_z, method = "spearman")
vldINS_L_169.eff_loc_cor <-
  graph_net_corr %>%
  ggplot(
    aes(
      x = vldINS_L_169.eff_loc, 
      y = siq_total_log
      )
    ) +
  geom_point() +
  geom_smooth(method = "lm") +
  theme_classic() +
  labs(
    x = "Ventral Granular INS (L) - local efficiency"
  ) +
  labeltheme
vldINS_L_169.eff_loc_cor 

# degree of the left Lateral superior frontal gyrus (lSFG_L_5.deg)
cor.test(graph_net_corr$siq_total_log, graph_net_corr$lSFG_L_5.deg)
cor.test(graph_net_corr$siq_total_log, graph_net_corr$lSFG_L_5.deg, method = "spearman")

lSFG_L_5.deg_cor <-
  graph_net_corr %>%
  ggplot(
    aes(
      x = lSFG_L_5.deg, 
      y = siq_total_log
      )
    ) +
  geom_point() +
  geom_smooth(method = "lm") +
  theme_classic() +
  labs(
    x = "Lateral SFG (L) - degree"
  ) +
  labeltheme
lSFG_L_5.deg_cor 

# degree of the Lateral posterior parahippocampal gyrus (R) (medPhG_R_120.deg)
cor.test(graph_net_corr$siq_total_log, graph_net_corr$medPhG_R_120.deg)
cor.test(graph_net_corr$siq_total_log, graph_net_corr$medPhG_R_120.deg, method = "spearman")
medPhG_R_120.deg_cor <-
  graph_net_corr %>%
  ggplot(
    aes(
      x = medPhG_R_120.deg, 
      y = siq_total_log
      )
    ) +
  geom_point() +
  geom_smooth(method = "lm") +
  theme_classic() +
  labs(
    x = "Medial PhG (R) - degree"
  ) +
  labeltheme
medPhG_R_120.deg_cor 

# eigenvector centrality of the left nucleus accumbens (NAC_L_227.eig_cent)
cor.test(graph_net_corr$siq_total_log, graph_net_corr$NAC_L_227.eig_cent)
cor.test(graph_net_corr$siq_total_log, graph_net_corr$NAC_L_227.eig_cent, method = "spearman")
NAC_L_227.eig_cent_cor <-
  graph_net_corr %>%
  ggplot(
    aes(
      x = NAC_L_227.eig_cent, 
      y = siq_total_log
      )
    ) +
  geom_point() +
  geom_smooth(method = "lm") +
  theme_classic() +
  labs(
    x = "NAcc (L) - eigenvector centrality"
  ) +
  labeltheme
NAC_L_227.eig_cent_cor 


brain_plus_all <-
  cowplot::plot_grid(
    internalizing_cor, 
    externalizing_cor, 
    lSFG_L_5.wm_z_cor, 
    rostMTG_R_84.pc_cor,
    postcentSPL_R_132.pc_cor, 
    operIFG_R_38.pc_cor, 
    caudMTG_L_81.deg_cor, 
    orbOrG_L_43.wm_z_cor, 
    vldINS_L_169.eff_loc_cor, 
    lSFG_L_5.deg_cor,
    medPhG_R_120.deg_cor,
    NAC_L_227.eig_cent_cor,
    label_size = 10,
    ncol = 3
    )
ggsave("brain_+_all_cors.png", brain_plus_all, width = 7, height = 7)
```

Are these 9 variables associated with mean FD?
```{r}
cor.test(graph_net_corr$siq_total_log, graph_net_corr$meanFD) # no corr (r = 0.04709477, p=0.6317)
cor.test(graph_net_corr$siq_total_log, graph_net_corr$meanFD, method = "spearman") # no corr (r = 0.04709477, p=0.6317)
cor.test(graph_net_corr$siq_total_log, graph_net_corr$meanGS)
# corr (r = 0.06121004, p=0.5331)
cor.test(graph_net_corr$siq_total_log, graph_net_corr$meanGS, method = "spearman")
# corr (r = 0.06121004, p=0.5331)
# T3_ysr_internalizing_sum, T3_ysr_externalizing_sum, lSFG_L_5.wm_z, rostMTG_R_84.pc, orbOrG_L_43.wm_z, vldINS_L_169.eff_loc, postcentSPL_R_132.pc, operIFG_R_38.pc
cor.test(graph_net_corr$T3_ysr_internalizing_sum, graph_net_corr$meanFD) # no corr
cor.test(graph_net_corr$T3_ysr_externalizing_sum, graph_net_corr$meanFD) # no corr
cor.test(graph_net_corr$lSFG_L_5.wm_z, graph_net_corr$meanFD) # no corr
cor.test(graph_net_corr$rostMTG_R_84.pc, graph_net_corr$meanFD) # no corr
cor.test(graph_net_corr$orbOrG_L_43.wm_z, graph_net_corr$meanFD) # no corr
cor.test(graph_net_corr$vldINS_L_169.eff_loc, graph_net_corr$meanFD) # sig correlation between local efficiency of the left ventrolateral insula (r=0.34, p=0.0004) and mean FD
cor.test(graph_net_corr$postcentSPL_R_132.pc, graph_net_corr$meanFD) # no corr
cor.test(graph_net_corr$operIFG_R_38.pc, graph_net_corr$meanFD) # no corr


summary(lm(scale(siq_total_log) ~ scale(vldINS_L_169.eff_loc) + scale(meanFD), data = graph_net_corr))
```

-------------------------------------------------------------------------
# Only Psychological and Environmental factors 
T1_ysr_internalizing_sum, T1_ysr_externalizing_sum, T3_ysr_internalizing_sum, T3_ysr_externalizing_sum,  
Parent_Education.T1, BMI.T1, sumsev_type_t1, sss_t1, sex_child, Age_at_T3Bx, Age_at_T1 interval,  
tanner_average.T3, tanner_average.T1
glimpse(graph_net_stand_siq_brain)

```{r}
graph_net_select_onlyPsych <-
  graph_net_stand_siq_brain %>%
  dplyr::select(
    -ends_with(".eff_loc"), 
    -ends_with(".deg"), 
   -ends_with(".eig_cent"),
    -ends_with("pc"), 
    -ends_with(".wm_z"), 
    -meanFD,
    -meanGS,
    -sleepy,
   -fmap_applied
    )
```

```{r}
#as.matrix
predictors_scaled_nonbrain <- data.matrix(graph_net_select_onlyPsych[2:17]) # columns 2 - 17 are predictors
siq_scaled_nonbrain <- graph_net_select_onlyPsych$siq_total_log_z # my response variable
save(predictors_scaled_nonbrain, file = "predictors_scaled_nonbrain.rda")
save(siq_scaled_nonbrain, file = "siq_scaled_nonbrain.rda")
```

## to get sense of coefficient path
```{r}
lasso_siq_nonbrain <- glmnet(predictors_scaled_nonbrain, siq_scaled_nonbrain, family = "gaussian", alpha = 1, standardize = FALSE)
# 
plot(lasso_siq_nonbrain, xvar = "dev", label = TRUE)
```
# nested cv
```{r}
n <- 106

## Your procedure

cvfit <- cv.glmnet(predictors_scaled_nonbrain, siq_scaled_nonbrain, standardize = FALSE, alpha = 1, nfolds = n)
betas <- coef(cvfit, s = "lambda.min")

# computing rsquared of unnested cv
# supp_siq_scaled is my y
X <- model.matrix(siq_total_log_z ~ ., data = graph_net_select_onlyPsych)[,-1]
# predict
y_hat <- predict(cvfit,  newx = X,  s = "lambda.min")
# R-Squared
r_sq <- 1 - (sum((siq_scaled_nonbrain - y_hat)^2) / sum((siq_scaled_nonbrain - mean(siq_scaled_nonbrain))^2))
r_sq # 0.59724 standardized prior to glmnet
cvfit$lambda.min # 0.04871511

## Get measure of performance of procedure
preds <- double(n)
ybars <- double(n)
for(i in 1:n){
    cvfit.inner <- cv.glmnet(predictors_scaled_nonbrain[-i,], siq_scaled_nonbrain[-i], standardize = FALSE, nfolds = n-1, grouped=FALSE, alpha=1)
    preds[i] <- predict(cvfit.inner, newx = predictors_scaled_nonbrain[i,,drop=FALSE], s = "lambda.min")
    ybars[i] <- mean(siq_scaled_nonbrain[-i])
    cat(i)
}

# total SS
mse0 <- mean((siq_scaled_nonbrain-ybars)^2) # 1.009524
# mean square error
mse <- mean((siq_scaled_nonbrain-preds)^2) # 0.4572163

R2 <-1- mse/mse0 # 0.5470971
```

# zero-order correlatoins

```{r}
graph_net_corr_onlypsych <- 
  graph_net_select_brain %>%
    dplyr::select(
    -ends_with(".eff_loc"), 
    -ends_with(".deg"), 
   -ends_with(".eig_cent"),
    -ends_with("pc"), 
    -ends_with(".wm_z"), 
    -meanFD,
    -meanGS,
    -sleepy,
   -fmap_applied,
   -siq_total.T3) %>%
  # converting race/ethinicity to factor and binarizing for analysis
  mutate(
    KSADS_Child_Race_by_P.T1 = factor(KSADS_Child_Race_by_P.T1),
    KSADS_Child_Race_by_P.T1_bin = ifelse(KSADS_Child_Race_by_P.T1 == "1", "1", "0"),
    # converting sex and sleepiness to factor
    sex_child = factor(sex_child),
    sex_child = ifelse(sex_child == "2", "1", "0")
    ) %>%
  dplyr::select(ELS_ID, siq_total_log, everything())
graph_net_corr_onlypsych <-
  graph_net_corr_onlypsych %>%
  mutate_at(vars(KSADS_Child_Race_by_P.T1_bin, sex_child), factor) %>%
  dplyr::select(-KSADS_Child_Race_by_P.T1)
contrasts(graph_net_corr_onlypsych$KSADS_Child_Race_by_P.T1_bin) = contr.treatment(2)
contrasts(graph_net_corr_onlypsych$sex_child) = contr.treatment(2)
```

```{r}
labeltheme <- theme(axis.title.y = element_blank(),
                    text = element_text(size = 7))
# concurrent internalizing severity
cor.test(graph_net_corr_onlypsych$siq_total_log, graph_net_corr_onlypsych$T3_ysr_internalizing_sum)
cor.test(graph_net_corr_onlypsych$siq_total_log, graph_net_corr_onlypsych$T3_ysr_internalizing_sum, method = "spearman")
internaliz_t3_cor <-
  graph_net_corr_onlypsych %>%
  ggplot(
    aes(
      x = T3_ysr_internalizing_sum, 
      y = siq_total_log
      )
    ) +
  geom_point() +
  geom_smooth(method = "lm") +
  theme_classic() +
  labs(
    x = "Concurrent Internalizing Severity"
  ) +
  labeltheme
internaliz_t3_cor

# concurrent externalizing severity
cor.test(graph_net_corr_onlypsych$siq_total_log, graph_net_corr_onlypsych$T3_ysr_externalizing_sum)
cor.test(graph_net_corr_onlypsych$siq_total_log, graph_net_corr_onlypsych$T3_ysr_externalizing_sum, method = "spearman")
extern_t3_cor <-
  graph_net_corr_onlypsych %>%
  ggplot(
    aes(
      x = T3_ysr_externalizing_sum, 
      y = siq_total_log
      )
    ) +
  geom_point() +
  geom_smooth(method = "lm") +
  theme_classic() +
  labs(
    x = "Concurrent Externalizing Severity"
  ) +
  labeltheme
extern_t3_cor

# internalizing at baseline
cor.test(graph_net_corr_onlypsych$siq_total_log, graph_net_corr_onlypsych$T1_ysr_internalizing_sum)
cor.test(graph_net_corr_onlypsych$siq_total_log, graph_net_corr_onlypsych$T1_ysr_internalizing_sum, method = "spearman")
intern_t1_cor <-
  graph_net_corr_onlypsych %>%
  ggplot(
    aes(
      x = T1_ysr_internalizing_sum, 
      y = siq_total_log
      )
    ) +
  geom_point() +
  geom_smooth(method = "lm") +
  theme_classic() +
  labs(
    x = "Baseline Internalizing Severity"
  ) +
  labeltheme
intern_t1_cor

# age at baseline
cor.test(graph_net_corr_onlypsych$siq_total_log, graph_net_corr_onlypsych$Age_at_T1)
cor.test(graph_net_corr_onlypsych$siq_total_log, graph_net_corr_onlypsych$Age_at_T1, method = "spearman")
age_t1_cor <-
  graph_net_corr_onlypsych %>%
  ggplot(
    aes(
      x = Age_at_T1, 
      y = siq_total_log
      )
    ) +
  geom_point() +
  geom_smooth(method = "lm") +
  theme_classic() +
  labs(
    x = "Baseline Age"
  ) +
  labeltheme
age_t1_cor

# baseline parent education 
cor.test(graph_net_corr_onlypsych$siq_total_log, graph_net_corr_onlypsych$Parent_Education.T1)
cor.test(graph_net_corr_onlypsych$siq_total_log, graph_net_corr_onlypsych$Parent_Education.T1, method = "spearman")
pe_cor <-
  graph_net_corr_onlypsych %>%
  ggplot(
    aes(
      x = Parent_Education.T1, 
      y = siq_total_log
      )
    ) +
  geom_point() +
  geom_smooth(method = "lm") +
  theme_classic() +
  labs(
    x = "Baseline Parent Education"
  ) +
  labeltheme
pe_cor

# concurrent bmi
cor.test(graph_net_corr_onlypsych$siq_total_log, graph_net_corr_onlypsych$BMI.T3)
cor.test(graph_net_corr_onlypsych$siq_total_log, graph_net_corr_onlypsych$BMI.T3, method = "spearman")
bmit3_cor <-
  graph_net_corr_onlypsych %>%
  ggplot(
    aes(
      x = BMI.T3, 
      y = siq_total_log
      )
    ) +
  geom_point() +
  geom_smooth(method = "lm") +
  theme_classic() +
  labs(
    x = "Concurrent BMI"
  ) +
  labeltheme
bmit3_cor


nonbrain <-
  cowplot::plot_grid(
internaliz_t3_cor, extern_t3_cor, intern_t1_cor, age_t1_cor, bmit3_cor, pe_cor,
    label_size = 10,
    ncol = 2
    )
ggsave("nonbrain_cors.png", nonbrain)
```


-------------------------------------------------------------------------


# Only Brain and brain related factors (global signal, meanfd, sleepiness)

## de-selecting variables
```{r}
graph_net_stand_siq_onlybrain <- graph_net_stand_siq_brain %>%
  dplyr::select(
    siq_total_log_z,
    ends_with(".eff_loc"), 
    ends_with(".deg"), 
    ends_with(".eig_cent"),
    ends_with("pc"), 
    ends_with(".wm_z"), 
    meanFD,
    meanGS,
    sleepy,
    fmap_applied
    ) %>%
  mutate(
    sleepy = factor(sleepy),
    fmap_applied = factor(fmap_applied)
  )
```


#### constructing x matrix and y vector
```{r}
#as.matrix
predictors_scaled_onlybrain <- data.matrix(graph_net_stand_siq_onlybrain[2:1255]) # columns 2 - 1254 are predictors
response_siq <-graph_net_stand_siq_onlybrain$siq_total_log_z # my response variable
save(predictors_scaled_onlybrain, file = "predictors_scaled_onlybrain.rda")
save(response_siq, file = "response_siq.rda")
```

## to get sense of coefficient path
```{r}
lasso_onlybrain <- glmnet(predictors_scaled_onlybrain, response_siq, family = "gaussian", alpha = 1, standardize = FALSE)
# 
plot(lasso_onlybrain, xvar = "dev", label = TRUE)
```

```{r}
n <- 106

## Your procedure

cvfit <- cv.glmnet(predictors_scaled_onlybrain, response_siq, standardize = FALSE, alpha = 1, nfolds = n)
betas <- coef(cvfit, s = "lambda.min")
# computing rsquared of unnested cv
# supp_siq_scaled is my y
X <- model.matrix(siq_total_log_z ~ ., data = graph_net_stand_siq_onlybrain)[,-1]
# predict
y_hat <- predict(cvfit,  newx = X,  s = "lambda.min")
# R-Squared
r_sq <- 1 - (sum((response_siq - y_hat)^2) / sum((response_siq - mean(response_siq))^2))
r_sq # 0.423974 standardized prior to glmnet
cvfit$lambda.min # 0.1602506


## Get measure of performance of procedure
preds <- double(n)
ybars <- double(n)
for(i in 1:n){
    cvfit.inner <- cv.glmnet(predictors_scaled_onlybrain[-i,], response_siq[-i], standardize = FALSE, nfolds = n-1, grouped=FALSE, alpha=1)
    preds[i] <- predict(cvfit.inner, newx = predictors_scaled_onlybrain[i,,drop=FALSE], s = "lambda.min")
    ybars[i] <- mean(response_siq[-i])
    cat(i)
}

mse0 <- mean((response_siq-ybars)^2) # 1.009524
mse <- mean((response_siq-preds)^2) # 1.178377

R2 <-1- mse/mse0 #-0.1672602
```

# zero order correlations
```{r}
graph_net_corr_onlybrain <- 
  graph_net_select_brain %>%
  # converting race/ethinicity to factor and binarizing for analysis
  dplyr::select(
    ELS_ID,
    siq_total_log,
    ends_with(".eff_loc"), 
    ends_with(".deg"), 
    ends_with(".eig_cent"),
    ends_with("pc"), 
    ends_with(".wm_z"), 
    meanFD,
    meanGS,
    sleepy,
    fmap_applied
    ) %>%
  mutate(
    sleepy = factor(sleepy),
    fmap_applied = factor(fmap_applied)
    ) %>%
  dplyr::select(ELS_ID, siq_total_log, everything())
graph_net_corr_onlybrain <-
  graph_net_corr_onlybrain %>%
  mutate_at(vars(sleepy, fmap_applied), factor)
contrasts(graph_net_corr_onlybrain$sleepy) = contr.treatment(2)
contrasts(graph_net_corr_onlybrain$fmap_applied) = contr.treatment(2)
```

```{r}
labeltheme <- theme(axis.title.y = element_blank(),
                    text = element_text(size = 7))
# within mod degree of the  lateral superior frontal gyrus (L) (lSFG_L_5.wm_z)
cor.test(graph_net_corr_onlybrain$siq_total_log, graph_net_corr_onlybrain$lSFG_L_5.wm_z)
lSFG_L_5.wm_z_cor <-
  graph_net_corr_onlybrain %>%
  ggplot(
    aes(
      x = lSFG_L_5.wm_z, 
      y = siq_total_log
      )
    ) +
  geom_point() +
  geom_smooth(method = "lm") +
  theme_classic() +
  labs(
    x = "Lateral SFG (L) - within-module degree"
  ) + 
  labeltheme
lSFG_L_5.wm_z_cor 


# participation coeff of the rostral MTG (rostMTG_R_84.pc)
cor.test(graph_net_corr_onlybrain$siq_total_log, graph_net_corr_onlybrain$rostMTG_R_84.pc)
rostMTG_R_84.pc_cor <-
  graph_net_corr_onlybrain %>%
  ggplot(
    aes(
      x = rostMTG_R_84.pc, 
      y = siq_total_log
      )
    ) +
  geom_point() +
  geom_smooth(method = "lm") +
  theme_classic() +
  labs(
    x = "Rostral MTG (R) - participation coefficient"
  ) + 
  labeltheme
rostMTG_R_84.pc_cor 

# participation coeff of the rostral MTG (rostMTG_R_84.pc)
cor.test(graph_net_corr_onlybrain$siq_total_log, graph_net_corr_onlybrain$rostMTG_R_84.pc)
rostMTG_R_84.pc_cor <-
  graph_net_corr_onlybrain %>%
  ggplot(
    aes(
      x = rostMTG_R_84.pc, 
      y = siq_total_log
      )
    ) +
  geom_point() +
  geom_smooth(method = "lm") +
  theme_classic() +
  labs(
    x = "Rostral MTG (R) - participation coefficient"
  ) + 
  labeltheme
rostMTG_R_84.pc_cor 

# participation coefficient of the right opercular area of the inferior frontal gyrus (operIFG_R_38.pc)
cor.test(graph_net_corr_onlybrain$siq_total_log, graph_net_corr_onlybrain$operIFG_R_38.pc)
operIFG_R_38.pc_cor <-
  graph_net_corr_onlybrain %>%
  ggplot(
    aes(
      x = operIFG_R_38.pc, 
      y = siq_total_log
      )
    ) +
  geom_point() +
  geom_smooth(method = "lm") +
  theme_classic() +
  labs(
    x = "IFG (R) - participation coefficient"
  ) +
  labeltheme
operIFG_R_38.pc_cor 

# within mod degree of the  medial superior frontal gyrus (R) (medSFG_R_14.wm_z)
cor.test(graph_net_corr_onlybrain$siq_total_log, graph_net_corr_onlybrain$medSFG_R_14.wm_z)
medSFG_R_14.wm_z_cor <-
  graph_net_corr_onlybrain %>%
  ggplot(
    aes(
      x = medSFG_R_14.wm_z, 
      y = siq_total_log
      )
    ) +
  geom_point() +
  geom_smooth(method = "lm") +
  theme_classic() +
  labs(
    x = "Medial SFG (R) - within-module degree"
  ) + 
  labeltheme
medSFG_R_14.wm_z_cor 

# local efficiency Lateral middle occipital gyrus (R) (mocgLOcC_R_200.eff_loc)
cor.test(graph_net_corr_onlybrain$siq_total_log, graph_net_corr_onlybrain$mocgLOcC_R_200.eff_loc)
mocgLOcC_R_200.eff_loc_cor <-
  graph_net_corr_onlybrain %>%
  ggplot(
    aes(
      x = mocgLOcC_R_200.eff_loc, 
      y = siq_total_log
      )
    ) +
  geom_point() +
  geom_smooth(method = "lm") +
  theme_classic() +
  labs(
    x = "Lateral mOccG (R) - local efficiency"
  ) +
  labeltheme
mocgLOcC_R_200.eff_loc_cor 


# within mod degree of theIntermediate ventral inferior temporal gyrus (R) (intvITG_R_90.wm_z)
cor.test(graph_net_corr_onlybrain$siq_total_log, graph_net_corr_onlybrain$intvITG_R_90.wm_z)
intvITG_R_90.wm_z_cor <-
  graph_net_corr_onlybrain %>%
  ggplot(
    aes(
      x = intvITG_R_90.wm_z, 
      y = siq_total_log
      )
    ) +
  geom_point() +
  geom_smooth(method = "lm") +
  theme_classic() +
  labs(
    x = "Ventral ITG (R) - within-module degree"
  ) + 
  labeltheme
intvITG_R_90.wm_z_cor

# Degree of the lateral ventral FuG (R) (latventFuG_R_108.deg)
cor.test(graph_net_corr_onlybrain$siq_total_log, graph_net_corr_onlybrain$latventFuG_R_108.deg)
latventFuG_R_108.deg_cor <-
  graph_net_corr_onlybrain %>%
  ggplot(
    aes(
      x = latventFuG_R_108.deg, 
      y = siq_total_log
      )
    ) +
  geom_point() +
  geom_smooth(method = "lm") +
  theme_classic() +
  labs(
    x = "Lateral Ventral FuG (R) - degree"
  ) + 
  labeltheme 
latventFuG_R_108.deg_cor

# participation coefficient of the right postcentral ssuperior parietal lobe (postcentSPL_R_132.pc)
cor.test(graph_net_corr_onlybrain$siq_total_log, graph_net_corr_onlybrain$postcentSPL_R_132.pc)
postcentSPL_R_132.pc_cor <-
  graph_net_corr_onlybrain %>%
  ggplot(
    aes(
      x = postcentSPL_R_132.pc, 
      y = siq_total_log
      )
    ) +
  geom_point() +
  geom_smooth(method = "lm") +
  theme_classic() +
  labs(
    x = "Postcentral SPL (R) - participation coefficient"
  ) +
  labeltheme
postcentSPL_R_132.pc_cor

# participation coeff of the rostral SPL (R) (rostSPL_R_126.pc)
cor.test(graph_net_corr_onlybrain$siq_total_log, graph_net_corr_onlybrain$rostSPL_R_126.pc)
rostSPL_R_126.pc_cor <-
  graph_net_corr_onlybrain %>%
  ggplot(
    aes(
      x = rostSPL_R_126.pc, 
      y = siq_total_log
      )
    ) +
  geom_point() +
  geom_smooth(method = "lm") +
  theme_classic() +
  labs(
    x = "Rostral SPL (R) - participation coefficient"
  ) + 
  labeltheme 
rostSPL_R_126.pc_cor


#local efficiency of the ventral granula insula gyrus (L) (vldINS_L_169.eff_loc)
cor.test(graph_net_corr_onlybrain$siq_total_log, graph_net_corr_onlybrain$vldINS_L_169.eff_loc)
vldINS_L_169.eff_loc_cor <-
  graph_net_corr_onlybrain %>%
  ggplot(
    aes(
      x = vldINS_L_169.eff_loc, 
      y = siq_total_log
      )
    ) +
  geom_point() +
  geom_smooth(method = "lm") +
  theme_classic() +
  labs(
    x = "Ventral Granular INS (L) - local efficiency"
  ) + 
  labeltheme 
vldINS_L_169.eff_loc_cor

# within mod degree of the dorsal MFG (L) (dMFG_L_15.wm_z)
cor.test(graph_net_corr_onlybrain$siq_total_log, graph_net_corr_onlybrain$dMFG_L_15.wm_z)
dMFG_L_15.wm_z_cor <-
  graph_net_corr_onlybrain %>%
  ggplot(
    aes(
      x = dMFG_L_15.wm_z, 
      y = siq_total_log
      )
    ) +
  geom_point() +
  geom_smooth(method = "lm") +
  theme_classic() +
  labs(
    x = "Dorsal MFG (L) - within module degree"
  ) + 
  labeltheme 
dMFG_L_15.wm_z_cor

# local efficiency of the medial STG (R) (medSTG_R_70.eff_loc)
cor.test(graph_net_corr_onlybrain$siq_total_log, graph_net_corr_onlybrain$medSTG_R_70.eff_loc)
medSTG_R_70.eff_loc_cor <-
  graph_net_corr_onlybrain %>%
  ggplot(
    aes(
      x = medSTG_R_70.eff_loc, 
      y = siq_total_log
      )
    ) +
  geom_point() +
  geom_smooth(method = "lm") +
  theme_classic() +
  labs(
    x = "Medial STG (R) - local efficiency"
  ) + 
  labeltheme
medSTG_R_70.eff_loc_cor


# local efficiency of theleft medial OrG (medOrG_L_47.eff_loc )
cor.test(graph_net_corr_onlybrain$siq_total_log, graph_net_corr_onlybrain$medOrG_L_47.eff_loc )
medOrG_L_47.eff_loc_cor <-
  graph_net_corr_onlybrain %>%
  ggplot(
    aes(
      x = medOrG_L_47.eff_loc, 
      y = siq_total_log
      )
    ) +
  geom_point() +
  geom_smooth(method = "lm") +
  theme_classic() +
  labs(
    x = "Medial orbital frontal gyrus (L) - local efficiency"
  ) + 
  labeltheme 
medOrG_L_47.eff_loc_cor

# within mod degree of the ventral IFG (L) (vIFG_L_39.wm_z)
cor.test(graph_net_corr_onlybrain$siq_total_log, graph_net_corr_onlybrain$vIFG_L_39.wm_z)
vIFG_L_39.wm_z_cor <-
  graph_net_corr_onlybrain %>%
  ggplot(
    aes(
      x = vIFG_L_39.wm_z, 
      y = siq_total_log
      )
    ) +
  geom_point() +
  geom_smooth(method = "lm") +
  theme_classic() +
  labs(
    x = "Ventral IFG (R) - within module degree"
  ) + 
  labeltheme 
vIFG_L_39.wm_z_cor


# local efficiency of the intermediate ventral inferior temporal gyrus (intvITG_R_90.eff_loc)
cor.test(graph_net_corr_onlybrain$siq_total_log, graph_net_corr_onlybrain$intvITG_R_90.eff_loc)
intvITG_R_90.eff_loc_cor <-
  graph_net_corr_onlybrain %>%
  ggplot(
    aes(
      x = intvITG_R_90.eff_loc, 
      y = siq_total_log
      )
    ) +
  geom_point() +
  geom_smooth(method = "lm") +
  theme_classic() +
  labs(
    x = "Ventral ITG (R) - local efficiency"
  ) + 
  labeltheme 
intvITG_R_90.eff_loc_cor

# within-mod degree of left orbital area 12/47 (orbOrG_L_43.wm_z)
cor.test(graph_net_corr_onlybrain$siq_total_log, graph_net_corr_onlybrain$orbOrG_L_43.wm_z)
orbOrG_L_43.wm_z_cor <-
  graph_net_corr_onlybrain %>%
  ggplot(
    aes(
      x = orbOrG_L_43.wm_z, 
      y = siq_total_log
      )
    ) +
  geom_point() +
  geom_smooth(method = "lm") +
  theme_classic() +
  labs(
    x = "OrG 12/47 (L) - within-module degree"
  ) +
  labeltheme
orbOrG_L_43.wm_z_cor 

# within mod degree of the lateral ventral fusiform gyrus (R) (latventFuG_R_108.wm_z)
cor.test(graph_net_corr_onlybrain$siq_total_log, graph_net_corr_onlybrain$latventFuG_R_108.wm_z)
latventFuG_R_108.wm_z_cor <-
  graph_net_corr_onlybrain %>%
  ggplot(
    aes(
      x = latventFuG_R_108.wm_z, 
      y = siq_total_log
      )
    ) +
  geom_point() +
  geom_smooth(method = "lm") +
  theme_classic() +
  labs(
    x = "Lateral ventral FuG (R) - within module degree"
  ) + 
  labeltheme 
latventFuG_R_108.wm_z_cor

# participation coeff of the medial PhG (medPhG_L_119.pc)
cor.test(graph_net_corr_onlybrain$siq_total_log, graph_net_corr_onlybrain$medPhG_L_119.pc)
medPhG_L_119.pc_cor <-
  graph_net_corr_onlybrain %>%
  ggplot(
    aes(
      x = medPhG_L_119.pc, 
      y = siq_total_log
      )
    ) +
  geom_point() +
  geom_smooth(method = "lm") +
  theme_classic() +
  labs(
    x = "Medial PhG (R) - participation coefficient"
  ) + 
  labeltheme 
medPhG_L_119.pc_cor

# Within mod degree of the temporal agranular insular PhG (R) (tempanginsPhG_R_118.wm_z)
cor.test(graph_net_corr_onlybrain$siq_total_log, graph_net_corr_onlybrain$tempanginsPhG_R_118.wm_z)
tempanginsPhG_R_118.wm_z_cor <-
  graph_net_corr_onlybrain %>%
  ggplot(
    aes(
      x = tempanginsPhG_R_118.wm_z, 
      y = siq_total_log
      )
    ) +
  geom_point() +
  geom_smooth(method = "lm") +
  theme_classic() +
  labs(
    x = "Temporal agranular insular PhG (R) - within module degree"
  ) + 
  labeltheme 
tempanginsPhG_R_118.wm_z_cor

#  eigenvec centrality of the intermediate lateral inferior temporal gyrus(intlatITG_R_96.eig_cent)
cor.test(graph_net_corr_onlybrain$siq_total_log, graph_net_corr_onlybrain$intlatITG_R_96.eig_cent)
intlatITG_R_96.eig_cent_cor <-
  graph_net_corr_onlybrain %>%
  ggplot(
    aes(
      x = intlatITG_R_96.eig_cent, 
      y = siq_total_log
      )
    ) +
  geom_point() +
  geom_smooth(method = "lm") +
  theme_classic() +
  labs(
    x = "Lateral ITG (R) - eigenvector centrality"
  ) + 
  labeltheme 
intlatITG_R_96.eig_cent_cor

#  eigenvec centrality of the rostral superior temporal gyrus (R) (rostSTG_R_80.eig_cent)
cor.test(graph_net_corr_onlybrain$siq_total_log, graph_net_corr_onlybrain$rostSTG_R_80.eig_cent)
rostSTG_R_80.eig_cent_cor <-
  graph_net_corr_onlybrain %>%
  ggplot(
    aes(
      x = rostSTG_R_80.eig_cent, 
      y = siq_total_log
      )
    ) +
  geom_point() +
  geom_smooth(method = "lm") +
  theme_classic() +
  labs(
    x = "Rostral STG (R) - eigenvector centrality"
  ) + 
  labeltheme 
rostSTG_R_80.eig_cent_cor

#  Within mod degree of the caudal MTG (L) (caudMTG_L_81.wm_z)
cor.test(graph_net_corr_onlybrain$siq_total_log, graph_net_corr_onlybrain$caudMTG_L_81.wm_z)
caudMTG_L_81.wm_z_cor <-
  graph_net_corr_onlybrain %>%
  ggplot(
    aes(
      x = caudMTG_L_81.wm_z, 
      y = siq_total_log
      )
    ) +
  geom_point() +
  geom_smooth(method = "lm") +
  theme_classic() +
  labs(
    x = "Caudal MTG (L) - within module degree"
  ) + 
  labeltheme 
caudMTG_L_81.wm_z_cor


brain <-
  cowplot::plot_grid(
    lSFG_L_5.wm_z_cor,
    rostMTG_R_84.pc_cor,
    operIFG_R_38.pc_cor, 
    medSFG_R_14.wm_z_cor,
    mocgLOcC_R_200.eff_loc_cor,    
    intvITG_R_90.wm_z_cor,
    latventFuG_R_108.deg_cor,    
    postcentSPL_R_132.pc_cor,
    rostSPL_R_126.pc_cor,    
    vldINS_L_169.eff_loc_cor,  
    dMFG_L_15.wm_z_cor,    
    medSTG_R_70.eff_loc_cor,    
    medOrG_L_47.eff_loc_cor,  
    vIFG_L_39.wm_z_cor,    
    intvITG_R_90.eff_loc_cor,    
    orbOrG_L_43.wm_z_cor,
    latventFuG_R_108.wm_z_cor,    
    medPhG_L_119.pc_cor,   
    tempanginsPhG_R_118.wm_z_cor,    
    intlatITG_R_96.eig_cent_cor,
    rostSTG_R_80.eig_cent_cor,
    caudMTG_L_81.wm_z_cor,
    label_size = 10,
    ncol = 3
    )
ggsave("brain_cors.png", brain, width = 7, height = 10)
```

#### just making sure I can interpret the brain predictors that were shared across the "all" and "brain only" model
```{r}
graph_net_stand_siq_identifiedIVs <- graph_net_stand_brain %>%
  dplyr::select(-ELS_ID, -siq_total_log) %>%
  dplyr::select(siq_total_log_z, lSFG_L_5.wm_z, rostMTG_R_84.pc, operIFG_R_38.pc, postcentSPL_R_132.pc, vldINS_L_169.eff_loc, orbOrG_L_43.wm_z) %>%
  dplyr::select(siq_total_log_z, everything()) # response (dv) variable in first column

#as.matrix
predictors_scaled_brain <- data.matrix(graph_net_stand_siq_identifiedIVs[2:7]) # columns 2 - 1271 are predictors
siq_scaled_brain <- graph_net_stand_siq_identifiedIVs$siq_total_log_z # my response variable


lasso_siq_brain <- glmnet(predictors_scaled_brain, siq_scaled_brain, family = "gaussian", alpha = 1, standardize = FALSE)
# 
plot(lasso_siq_brain, xvar = "dev", label = TRUE)

library(remotes)
n <- 106

## Your procedure

cvfit <- cv.glmnet(predictors_scaled_brain, siq_scaled_brain, standardize = FALSE, alpha = 1, nfolds = n)
betas <- coef(cvfit, s = "lambda.min")

# computing rsquared of unnested cv
# supp_siq_scaled is my y
X <- model.matrix(siq_total_log_z ~ ., data = graph_net_stand_siq_identifiedIVs)[,-1]
# predict
y_hat <- predict(cvfit,  newx = X,  s = "lambda.min")
# R-Squared
r_sq <- 1 - (sum((siq_scaled_brain - y_hat)^2) / sum((siq_scaled_brain - mean(siq_scaled_brain))^2))
r_sq # 0.3804921 standardized prior to glmnet
cvfit$lambda.min # 0.002118419

## Get measure of performance of procedure
preds <- double(n)
ybars <- double(n)
for(i in 1:n){
    cvfit.inner <- cv.glmnet(predictors_scaled_brain[-i,], siq_scaled_brain[-i], standardize = FALSE, nfolds = n-1, grouped=FALSE, alpha=1)
    preds[i] <- predict(cvfit.inner, newx = predictors_scaled_brain[i,,drop=FALSE], s = "lambda.min")
    ybars[i] <- mean(siq_scaled_brain[-i])
    cat(i)
}

# total SS
mse0 <- mean((siq_scaled_brain-ybars)^2) # 1.009524
# mean square error
mse <- mean((siq_scaled_brain-preds)^2) # 0.7123318

R2 <-1- mse/mse0 # 0.2943883
```


## seeing whether it's the drop in participant or the non-SDC that yields the nofmaps results
# reading in data
```{r, message=FALSE}
graph_net_fp <- "~/Box/Mooddata_Coordinating/1_Lab_Coordinating/Users/JackieSchwartz/PHIND_RS_and_Suicide/April2021Analysis/fmaps_and_nofmaps/siq_graph.csv"

graph_net <- read_csv(graph_net_fp)
```


# function to z-score
```{r}
options(scipen=999) 
z_score <- function(x) {
diff_mu <- x - mean(x, na.rm = T)
sd <- sd(x, na.rm = T)
diff_mu / sd
}
```


## de-selecting variables
```{r}
graph_net_select_brain_105 <- graph_net %>%
  dplyr::select(
    -c(meds_baseline)
    ) %>%
  filter(!ELS_ID=="134")
# already filtered out those with history of stbs
```

#### recoding and standardizing
```{r}
graph_net_stand_brain_105 <- graph_net_select_brain_105 %>%
  # converting race/ethinicity to factor and binarizing for analysis
  mutate(
    KSADS_Child_Race_by_P.T1 = factor(KSADS_Child_Race_by_P.T1),
    KSADS_Child_Race_by_P.T1_bin = ifelse(KSADS_Child_Race_by_P.T1 == "1", "1", "0"),
    # converting sex and sleepiness to factor
    sex_child = factor(sex_child),
    sex_child = ifelse(sex_child == "2", "1", "0"),
    sleepy = factor(sleepy),
    fmap_applied = factor(fmap_applied)
    ) %>%
  dplyr::select(-KSADS_Child_Race_by_P.T1, -siq_total.T3) %>%
  # z-scoring all numeric variables
  mutate_at(vars(-ELS_ID, -siq_total_log, -KSADS_Child_Race_by_P.T1_bin, -sex_child, -sleepy, -fmap_applied), z_score) %>%
  # arranging with ELS_ID first, then siq, then everything else
  mutate(siq_total_log_z = z_score(siq_total_log)) %>%
  dplyr::select(ELS_ID, siq_total_log_z, siq_total_log, everything())
```

#### setting contrasts
```{r}
graph_net_stand_brain_105 <-
  graph_net_stand_brain_105 %>%
  mutate_at(vars(KSADS_Child_Race_by_P.T1_bin, sex_child, sleepy, fmap_applied), factor)
contrasts(graph_net_stand_brain_105$KSADS_Child_Race_by_P.T1_bin) = contr.treatment(2)
contrasts(graph_net_stand_brain_105$sex_child) = contr.treatment(2)
contrasts(graph_net_stand_brain_105$sleepy) = contr.treatment(2)
contrasts(graph_net_stand_brain_105$fmap_applied) = contr.treatment(2)
```

#### constructing x matrix and y vector
```{r}
graph_net_stand_siq_brain_105 <- graph_net_stand_brain_105 %>%
  dplyr::select(-ELS_ID, -siq_total_log) %>%
  dplyr::select(siq_total_log_z, everything()) # response (dv) variable in first column

#as.matrix
predictors_scaled_brain <- data.matrix(graph_net_stand_siq_brain_105[2:1271]) # columns 2 - 1271 are predictors
siq_scaled_brain <- graph_net_stand_siq_brain_105$siq_total_log_z # my response variable

```

## to get sense of coefficient path
```{r}
lasso_siq_brain <- glmnet(predictors_scaled_brain, siq_scaled_brain, family = "gaussian", alpha = 1, standardize = FALSE)
# 
plot(lasso_siq_brain, xvar = "dev", label = TRUE)
```
# nested cv
```{r}
library(remotes)
n <- 105

## Your procedure

cvfit <- cv.glmnet(predictors_scaled_brain, siq_scaled_brain, standardize = FALSE, alpha = 1, nfolds = n)
betas <- coef(cvfit, s = "lambda.min")

# computing rsquared of unnested cv
# supp_siq_scaled is my y
X <- model.matrix(siq_total_log_z ~ ., data = graph_net_stand_siq_brain_105)[,-1]
# predict
y_hat <- predict(cvfit,  newx = X,  s = "lambda.min")
# R-Squared
r_sq <- 1 - (sum((siq_scaled_brain - y_hat)^2) / sum((siq_scaled_brain - mean(siq_scaled_brain))^2))
r_sq # 0.6708242 standardized prior to glmnet
cvfit$lambda.min # 0.135493

## Get measure of performance of procedure
preds <- double(n)
ybars <- double(n)
for(i in 1:n){
    cvfit.inner <- cv.glmnet(predictors_scaled_brain[-i,], siq_scaled_brain[-i], standardize = FALSE, nfolds = n-1, grouped=FALSE, alpha=1)
    preds[i] <- predict(cvfit.inner, newx = predictors_scaled_brain[i,,drop=FALSE], s = "lambda.min")
    ybars[i] <- mean(siq_scaled_brain[-i])
    cat(i)
}

# total SS
mse0 <- mean((siq_scaled_brain-ybars)^2) # 1.009615
# mean square error
mse <- mean((siq_scaled_brain-preds)^2) # 0.4702881

R2 <-1- mse/mse0 # 0.5341908
```
